+++ src/vue/Section/Settings.vue
@@ -323,8 +323,10 @@
                     });
             },
             resetSettingsAction() {
-                Messages.confirm('This will reset all settings to their defaults. Do you want to continue?',
-                                 'Reset all settings')
+                Messages.confirm(
+                    'This will reset all settings to their defaults. Do you want to continue?',
+                                 'Reset all settings'
+                )
                     .then(() => { this.resetSettings(); });
             },
             async resetSettings() {
@@ -382,6 +384,9 @@
             },
             advanced(value) {
                 this.advancedSettings = this.settings['client.settings.advanced'] = value === '1';
+            },
+            locked(value) {
+                document.getElementById('app-content').classList.toggle('blocking');
             }
         }
     };